vet-edit-report - move report-card into vet/_lib
